Refactor into components and rename favs service objects.
