+++ backend/spec/fixtures/vcr_cassettes/Favorites/POST_/api/v1/favorites/behaves_like_JWT_validation/when_the_JWT_is_valid/when_the_user_exists/success.yml
@@ -1,0 +1,51 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.thecatapi.com/v1/favourites
+    body:
+      encoding: UTF-8
+      string: '{"image_id":"YTvbh1DEh","sub_id":"example@gmail.com"}'
+    headers:
+      User-Agent:
+      - Faraday v1.7.2
+      Content-type:
+      - application/json
+      x-api-key:
+      - token [CATS_API_KEY]
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      x-dns-prefetch-control:
+      - 'off'
+      x-frame-options:
+      - SAMEORIGIN
+      strict-transport-security:
+      - max-age=15552000; includeSubDomains
+      x-download-options:
+      - noopen
+      x-content-type-options:
+      - nosniff
+      x-xss-protection:
+      - 1; mode=block
+      vary:
+      - Origin
+      content-type:
+      - application/json; charset=utf-8
+      x-response-time:
+      - 135ms
+      x-cloud-trace-context:
+      - e744474593b1ea2a52544844f15a8e7f
+      date:
+      - Sun, 31 Oct 2021 17:54:32 GMT
+      server:
+      - Google Frontend
+      content-length:
+      - '34'
+    body:
+      encoding: UTF-8
+      string: '{"message":"SUCCESS","id":2138616}'
+  recorded_at: Sun, 31 Oct 2021 17:54:32 GMT
+recorded_with: VCR 6.0.0
