--- backend/spec/fixtures/vcr_cassettes/Favourites/GET_/api/v1/favourites/success/returns_the_favs.yml
@@ -1,57 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://api.thecatapi.com/v1/favourites?limit=3&order=Asc&page=1
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      x-api-key:
-      - token [CATS_API_KEY]
-      User-Agent:
-      - Faraday v1.7.2
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      x-dns-prefetch-control:
-      - 'off'
-      x-frame-options:
-      - SAMEORIGIN
-      strict-transport-security:
-      - max-age=15552000; includeSubDomains
-      x-download-options:
-      - noopen
-      x-content-type-options:
-      - nosniff
-      x-xss-protection:
-      - 1; mode=block
-      vary:
-      - Origin
-      pagination-count:
-      - '6'
-      pagination-page:
-      - '1'
-      pagination-limit:
-      - '3'
-      access-control-expose-headers:
-      - Pagination-Count, Pagination-Page, Pagination-Limit
-      content-type:
-      - application/json; charset=utf-8
-      x-response-time:
-      - 127ms
-      x-cloud-trace-context:
-      - 12dace900a4b6454bda590420b46b2b5
-      date:
-      - Fri, 29 Oct 2021 23:30:14 GMT
-      server:
-      - Google Frontend
-      content-length:
-      - '521'
-    body:
-      encoding: UTF-8
-      string: '[{"id":2102245,"user_id":"dccq5n","image_id":"v","sub_id":null,"created_at":"2021-09-19T22:15:26.000Z","image":{"id":"v","url":"https://cdn2.thecatapi.com/images/v.jpg"}},{"id":2102473,"user_id":"dccq5n","image_id":"1","sub_id":"challenge","created_at":"2021-09-20T14:49:03.000Z","image":{"id":"1","url":"https://cdn2.thecatapi.com/images/1.jpg"}},{"id":2102486,"user_id":"dccq5n","image_id":"1c","sub_id":null,"created_at":"2021-09-20T16:53:05.000Z","image":{"id":"1c","url":"https://cdn2.thecatapi.com/images/1c.gif"}}]'
-  recorded_at: Fri, 29 Oct 2021 23:30:13 GMT
-recorded_with: VCR 6.0.0
