--- backend/spec/fixtures/vcr_cassettes/Favourites/POST_/api/v1/favourites/behaves_like_JWT_validation/when_the_JWT_is_valid/when_the_user_exists/success.yml
@@ -1,50 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.thecatapi.com/v1/favourites
-    body:
-      encoding: UTF-8
-      string: '{"image_id":"g","sub_id":"example@gmail.com"}'
-    headers:
-      User-Agent:
-      - Faraday v1.7.2
-      Content-type:
-      - application/json
-      X-api-key:
-      - 51661707-c326-49f5-9da7-47588969d272
-  response:
-    status:
-      code: 400
-      message: Bad Request
-    headers:
-      x-dns-prefetch-control:
-      - 'off'
-      x-frame-options:
-      - SAMEORIGIN
-      strict-transport-security:
-      - max-age=15552000; includeSubDomains
-      x-download-options:
-      - noopen
-      x-content-type-options:
-      - nosniff
-      x-xss-protection:
-      - 1; mode=block
-      vary:
-      - Origin
-      content-type:
-      - application/json; charset=utf-8
-      x-cloud-trace-context:
-      - 199547ab69f8e0d4571125adbede63ae
-      date:
-      - Fri, 29 Oct 2021 23:58:34 GMT
-      server:
-      - Google Frontend
-      content-length:
-      - '117'
-    body:
-      encoding: UTF-8
-      string: '{"message":"DUPLICATE_FAVOURITE - favourites are unique for account
-        + image_id + sub_id","status":400,"level":"info"}'
-  recorded_at: Fri, 29 Oct 2021 23:58:34 GMT
-recorded_with: VCR 6.0.0
