--- backend/spec/fixtures/vcr_cassettes/Favourites/POST_/api/v1/favourites/success/returns_the_fav_id.yml
@@ -1,51 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://api.thecatapi.com/v1/favourites
-    body:
-      encoding: UTF-8
-      string: '{"image_id":"g","sub_id":"example@gmail.com"}'
-    headers:
-      User-Agent:
-      - Faraday v1.7.2
-      Content-type:
-      - application/json
-      X-api-key:
-      - 51661707-c326-49f5-9da7-47588969d272
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      x-dns-prefetch-control:
-      - 'off'
-      x-frame-options:
-      - SAMEORIGIN
-      strict-transport-security:
-      - max-age=15552000; includeSubDomains
-      x-download-options:
-      - noopen
-      x-content-type-options:
-      - nosniff
-      x-xss-protection:
-      - 1; mode=block
-      vary:
-      - Origin
-      content-type:
-      - application/json; charset=utf-8
-      x-response-time:
-      - 142ms
-      x-cloud-trace-context:
-      - 37dc302da579420ca3391d0dbb7948e7
-      date:
-      - Fri, 29 Oct 2021 23:58:05 GMT
-      server:
-      - Google Frontend
-      content-length:
-      - '34'
-    body:
-      encoding: UTF-8
-      string: '{"message":"SUCCESS","id":2138537}'
-  recorded_at: Fri, 29 Oct 2021 23:58:05 GMT
-recorded_with: VCR 6.0.0
